% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{format_for_recode}
\alias{format_for_recode}
\title{Format for recode}
\usage{
format_for_recode(target, as_mutate = FALSE)
}
\arguments{
\item{target}{The factor you want to recode}

\item{as_mutate}{Output as a mutate call. Defaults to false.}
}
\description{
Print an incomplete fct_recode function to make it easier to manually recode factor levels.
}
