% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{format_for_rename}
\alias{format_for_rename}
\title{Format for rename}
\usage{
format_for_rename(data)
}
\arguments{
\item{data}{Dataframe to rename}
}
\description{
Prints an incomplete rename function to make it easy to manually rename data frame columns.
}
