% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{plot_japanese_text}
\alias{plot_japanese_text}
\title{Plot Japanese text}
\usage{
plot_japanese_text()
}
\description{
Update current ggplot theme to use a font that supports Japanese text.
}
\details{
Note: run `toolbelt::plot_japanese_text` to get the source to paste into a
workbook.
}
